Move JSW specifics (well most of them) into bootstrap.jsw package
