+++ nexus/nexus-bootstrap/src/main/java/org/sonatype/nexus/bootstrap/CommandMonitorThread.java
@@ -23,6 +23,8 @@ import java.net.InetAddress;
 import java.net.ServerSocket;
 import java.net.Socket;
 
+// FIXME: This is presently JSW specific
+
 /**
  * Thread which listens for command messages to control the JVM.
  *
