+++ nexus/nexus-bootstrap/src/main/java/org/sonatype/nexus/bootstrap/jsw/WrapperLog.java
@@ -10,7 +10,7 @@
  * of Sonatype, Inc. Apache Maven is a trademark of the Apache Software Foundation. M2eclipse is a trademark of the
  * Eclipse Foundation. All other trademarks are the property of their respective owners.
  */
-package org.sonatype.nexus.bootstrap;
+package org.sonatype.nexus.bootstrap.jsw;
 
 import org.tanukisoftware.wrapper.WrapperManager;
