+++ nexus/nexus-oss-webapp/src/main/resources/content/bin/jsw/conf/wrapper.conf
@@ -15,7 +15,7 @@ wrapper.working.dir=../../..
 wrapper.java.command=java
 
 # The main class that JSW will execute within JVM
-wrapper.java.mainclass=org.sonatype.nexus.bootstrap.JswLauncher
+wrapper.java.mainclass=org.sonatype.nexus.bootstrap.jsw.JswLauncher
 
 # The JVM classpath
 wrapper.java.classpath.1=bin/jsw/lib/wrapper-3.2.3.jar
